seperate test for tx circuit
